app/requirements.txt
boto3==1.35.7
fastapi==0.112.2
llama-index==0.11.3
llama-index-embeddings-openai==0.2.3
llama-index-vector-stores-qdrant==0.3.0
qdrant-client==1.11.1
python-dotenv==1.0.1
uvicorn==0.30.6
openai==1.43.0